extraction_engine/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
bce-python-sdk==0.9.55
beautifulsoup4==4.14.3
cachetools==6.2.2
camelot-py==1.0.9
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
effdet==0.4.1
et_xmlfile==2.0.0
filelock==3.20.0
fsspec==2025.12.0
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-documentai==3.7.0
google-genai==1.53.0
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
imagesize==1.4.1
img2table==1.4.2
iopath==0.1.10
Jinja2==3.1.6
jiter==0.12.0
layoutparser==0.3.4
Levenshtein==0.27.3
llvmlite==0.45.1
MarkupSafe==3.0.3
modelscope==1.32.0
mpmath==1.3.0
networkx==3.6
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.10
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
prettytable==3.17.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
thefuzz==0.22.1
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
xlsxwriter==3.2.9
